Refactor lz4cust code to already allow for non-CGO builds
