+++ pkg/goDB/encoder/encoder_test.go
@@ -2,11 +2,13 @@ package encoder
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 	"testing"
 
 	"github.com/els0r/goProbe/pkg/goDB/encoder/encoders"
 	"github.com/els0r/goProbe/pkg/goDB/encoder/lz4"
+	"github.com/els0r/goProbe/pkg/goDB/encoder/lz4cust"
 	"github.com/els0r/goProbe/pkg/goDB/encoder/zstd"
 )
 
@@ -27,7 +29,7 @@ func TestNewByString(t *testing.T) {
 		{"empty string", "", encoders.EncoderTypeNull, false},
 		{"null encoder", "null", encoders.EncoderTypeNull, false},
 		{"lz4 encoder", "lz4", encoders.EncoderTypeLZ4, false},
-		{"lz4 encoder custom", "lz4cust", encoders.EncoderTypeLZ4Custom, false},
+		{"lz4 encoder custom", "lz4cust", encoders.EncoderTypeLZ4Custom, lz4cust.ErrDeprecated != nil},
 		{"lz4 encoder (uppercase)", "LZ4", encoders.EncoderTypeLZ4, false},
 		{"zstd encoder", "zstd", encoders.EncoderTypeZSTD, false},
 		{"zstd encoder (uppercase)", "ZSTD", encoders.EncoderTypeZSTD, false},
@@ -61,6 +63,9 @@ func TestCompressionDecompression(t *testing.T) {
 		t.Run(encType.String(), func(t *testing.T) {
 			enc, err := New(encType)
 			if err != nil {
+				if errors.Is(err, lz4cust.ErrDeprecated) {
+					t.Skip("ignoring deprecated encoding algorithm")
+				}
 				t.Fatalf("Failed to instantiate encoder of type %s: %s", encType, err)
 			}
 			defer func() {
@@ -102,6 +107,9 @@ func TestCompressionDecompressionCustomLevel(t *testing.T) {
 			t.Run(fmt.Sprintf("%s_%d", encType, level), func(t *testing.T) {
 				enc, err := New(encType)
 				if err != nil {
+					if errors.Is(err, lz4cust.ErrDeprecated) {
+						t.Skip("ignoring deprecated encoding algorithm")
+					}
 					t.Fatalf("Failed to instantiate encoder of type %s: %s", encType, err)
 				}
 				enc.SetLevel(level)
@@ -144,6 +152,9 @@ func BenchmarkEncodersCompress(b *testing.B) {
 		b.Run(encType.String(), func(b *testing.B) {
 			enc, err := New(encType)
 			if err != nil {
+				if errors.Is(err, lz4cust.ErrDeprecated) {
+					b.Skip("ignoring deprecated encoding algorithm")
+				}
 				b.Fatalf("Failed to instantiate encoder of type %s: %s", encType, err)
 			}
 
@@ -175,6 +186,9 @@ func BenchmarkEncodersDecompress(b *testing.B) {
 		b.Run(encType.String(), func(b *testing.B) {
 			enc, err := New(encType)
 			if err != nil {
+				if errors.Is(err, lz4cust.ErrDeprecated) {
+					b.Skip("ignoring deprecated encoding algorithm")
+				}
 				b.Fatalf("Failed to instantiate encoder of type %s: %s", encType, err)
 			}
 			defer func() {
